[1/n] More clang -Wconversion fixes
Similar to gcc. Also for clang we now have a larger number of classes
that each only produce 1 or a few warnings. We fix those in this and the
next few patches.

+++ src/ide/MB89352.cc
@@ -596,7 +596,7 @@ void MB89352::writeRegister(uint8_t reg, uint8_t value)
 		// set Bus Device ID
 		value &= 7;
 		myId = value;
-		regs[REG_BDID] = 1 << value;
+		regs[REG_BDID] = uint8_t(1 << value);
 		break;
 
 		// Nothing
