+++ src/memory/EEPROM_93C46.cc
@@ -120,7 +120,7 @@ void EEPROM_93C46::clockEvent(EmuTime::param time)
 		break;
 
 	case WAIT_FOR_COMMAND:
-		shiftRegister = (shiftRegister << 1) | int(pinDI);
+		shiftRegister = narrow_cast<uint16_t>((shiftRegister << 1) | int(pinDI));
 		++bits;
 		if (bits == (2 + ADDRESS_BITS)) {
 			execute_command(time);
@@ -130,7 +130,7 @@ void EEPROM_93C46::clockEvent(EmuTime::param time)
 	case READING_DATA:
 		if ((bits % DATA_BITS) == 0) {
 			uint8_t value = read(address);
-			shiftRegister = value << (SHIFT_REG_BITS - DATA_BITS);
+			shiftRegister = uint16_t(value << (SHIFT_REG_BITS - DATA_BITS));
 			address = (address + 1) & ADDRESS_MASK;
 		} else {
 			shiftRegister <<= 1;
@@ -139,7 +139,7 @@ void EEPROM_93C46::clockEvent(EmuTime::param time)
 		break;
 
 	case WAIT_FOR_WRITE:
-		shiftRegister = (shiftRegister << 1) | int(pinDI);
+		shiftRegister = narrow_cast<uint16_t>((shiftRegister << 1) | int(pinDI));
 		++bits;
 		if (bits == DATA_BITS) {
 			if (writeProtected) {
@@ -152,7 +152,7 @@ void EEPROM_93C46::clockEvent(EmuTime::param time)
 		break;
 
 	case WAIT_FOR_WRITE_ALL:
-		shiftRegister = (shiftRegister << 1) | int(pinDI);
+		shiftRegister = narrow_cast<uint16_t>((shiftRegister << 1) | int(pinDI));
 		++bits;
 		if (bits == DATA_BITS) {
 			if (writeProtected) {
